requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
streamlit==1.37.0
requests==2.32.3
